requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
resend
python-dotenv
cloudinary